requirements.txt
Flask
Flask-SQLAlchemy
pymysql
flask-session
python-dotenv
flask-bcrypt
requests
Flask-WTF